fixup! nfc: Revert oplus changes
Yeah, if we're gonna drop all usage of the oplus
functions, and it's not defined in the defconfig,
don't make it build by default anyway.

Newer versions of clang or something get overly
enthusiastic and try to do it anyway, and we do
not have the dependencies for that business yet.